globus_configuration/requirements.txt
globus-compute-endpoint
globus-compute-sdk
globus-sdk
python-dotenv
